+++ core/tailcat.py
@@ -25,7 +25,7 @@ try:
 except ImportError:
     from ingress import GuardServer, Handler, watchdog
 
-VERSION = "6.2.2"
+VERSION = "6.3.1"
 DEFAULT_HOST = "0.0.0.0"
 DEFAULT_PORT = 7443
 DEFAULT_BACKEND_HOST = "127.0.0.1"